Workload division improved

+++ include/chunk.h
@@ -9,13 +9,13 @@ MPI_Datatype mpi_text_file_chunk;
 
 typedef struct chunk {
   char filename[FILENAME_SIZE]; // può essere migliorato con una allocazione dinamica
-  size_t start_index;
-  size_t end_index;
+  int start_index;
+  int end_index;
 }chunk;
 
 void initialize_chunk_type();
 
-chunk new_chunk(const char filename[], size_t start_index,size_t end_index);
+chunk new_chunk(const char filename[], int start_index, int end_index);
 
 bool chunk_equals(chunk c1,chunk c2);
